+++ test/cases/loader-option-preprocessor-return-null/webpack.config.js
@@ -1,0 +1,49 @@
+const path = require('path');
+const HtmlBundlerPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+    clean: true,
+  },
+
+  plugins: [
+    new HtmlBundlerPlugin({
+      entry: {
+        index: {
+          import: './src/index.html',
+          data: {
+            title: 'Home',
+            headline: 'Breaking Bad',
+            name: {
+              firstname: 'Walter',
+              lastname: 'White',
+            },
+          },
+        },
+      },
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.(html)$/,
+        loader: HtmlBundlerPlugin.loader,
+        options: {
+          // test preprocessor return null, has same effect as `preprocessor: false`
+          preprocessor: () => null,
+        },
+      },
+      {
+        test: /\.(png|svg|jpe?g|webp)$/i,
+        type: 'asset/resource',
+        generator: {
+          filename: 'assets/img/[name].[hash:8][ext]',
+        },
+      },
+    ],
+  },
+};